filter configs
adds a new serializable filter schema, FilterConfig, plus a set of
filter functions to transform a FilterConfig into a task predicate.

the existing FilterControls UI isn't compatible with the new config
schema, so instead we add a handful of presets (defaultConfigs) and a
picker to choose between them on the task list.

the end goal is to make FilterConfigs fully customizable and stored in
redux state.

also updates `calcUrgency` to return 0 if the task was already
completed, to make the unfiltered task list push those to the bottom.