navigation refactor
still very wip but the bones are there and it works!

this commit pulls in drawer and bottom-tabs navigation. the new
structure looks like this:

stack nav
|
|- drawer nav
|  |- main page (multiTaskView)
|  |  |- filter controls
|  |  +- bottom tab nav
|  |    |- task list
|  |    +- multi task metrics
|  |
|  +- settings, about pages, etc
|
|- singleTaskView / another bottom tab nav
|  |- task details (formerly ViewTask)
|  +- singleTaskMetrics
|
+- modals: create task, edit task

note: it might be better to move the single task view inside the drawer
so you can open the drawer while you're on it without having to hit back
first