remove middleware, add click effects
trying to make the task list faster by removing the immutability
and serializability checking middleware (these are disabled in prod
builds, but this app is permanently in development mode).
also turned off detaching the task list when viewing a task, so it
doesn't have to be loaded again.

this commit also adds the android ripple to TaskListItem (much needed)
and Button. this does NOT affect IconButton unfortunately. the ripple on
that isn't playing well with the border radius (it's square on round
buttons). there's also some weird junk code about styling based on
`pressed` state - this was used for the buttons in the old FilterControl
component. it is bad and needs to be refactored.